adminer_backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dateparser==1.0.0
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.3
django-admin-rangefilter==0.8.1
django-anymail==8.3
django-bootstrap-modal-forms==2.2.0
django-bootstrap-v5==1.0.4
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-import-export==2.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
Faker==8.4.0
filelock==3.0.12
flake8==3.9.2
flake8-builtins==1.5.3
flake8-print==4.0.0
greenlet==1.1.0
idna==2.10
importlib-metadata==3.10.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jet-bridge-base==1.0.1
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.5.29
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-core==4.1.0
soupsieve==2.2.1
SQLAlchemy==1.4.18
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
